refactor fitgauss to standalone code
