+++ ngmix/metacal/__init__.py
@@ -6,3 +6,5 @@ from .defaults import *
 from .convenience import *
 from . import gauss_target_psf
 from . gauss_target_psf import get_gauss_target_psf
+from . import fitgauss_target_psf
+from . fitgauss_target_psf import get_fitgauss_target_psf
